+++ examples/julia.rs
@@ -1,27 +1,30 @@
 extern crate minifb;
 
-use minifb::{Key, Window, WindowOptions};
+use minifb::{Key, Scale, ScaleMode, Window, WindowOptions};
 
-const WIDTH: usize = 600;
-const HEIGHT: usize = 600;
+const WIDTH: usize = 100;
+const HEIGHT: usize = 100;
 const FRACTAL_DEPTH: u32 = 64;
 const GENERATION_INFINITY: f64 = 16.;
 
 fn main() {
     let mut buffer: Vec<u32> = vec![0; WIDTH * HEIGHT];
 
-    let mut window = match Window::new(
+    let mut window = Window::new(
         "Fractal - ESC to exit",
         WIDTH,
         HEIGHT,
-        WindowOptions::default(),
-    ) {
-        Ok(win) => win,
-        Err(err) => {
-            println!("Unable to create window {}", err);
-            return;
-        }
-    };
+        WindowOptions {
+            resize: true,
+            scale: Scale::X2,
+            scale_mode: ScaleMode::AspectRatioStretch,
+            ..WindowOptions::default()
+        },
+    )
+    .expect("Unable to Open Window");
+
+    // Limit to max ~60 fps update rate
+    window.limit_update_rate(Some(std::time::Duration::from_micros(16600)));
 
     let range = 2.0;
     let x_min = 0. - range;
@@ -31,6 +34,8 @@ fn main() {
     let y_max = 0. + range;
 
     let mut angle: f64 = 0.0;
+
+    window.set_background_color(0, 0, 20);
 
     while window.is_open() && !window.is_key_down(Key::Escape) {
         for i in 0..buffer.len() {
@@ -58,7 +63,7 @@ fn main() {
         angle += 0.1;
 
         // We unwrap here as we want this code to exit if it fails
-        window.update_with_buffer_size(&buffer, WIDTH, HEIGHT).unwrap();
+        window.update_with_buffer(&buffer, WIDTH, HEIGHT).unwrap();
     }
 }